amélioration de la détection du network + ImageView a la place de backgroundImage

+++ app/controllers/index.js
@@ -24,8 +24,12 @@ activityIndicator.show();
 
 /*for test */
 //////////////////////////////////////
-happyhour.deleteAll();
-etablishment.deleteAll();
+if(Alloy.Globals.hasConnection) {
+    Ti.API.info('has connection');
+    happyhour.deleteAll();
+    etablishment.deleteAll();
+}
+
 
 
 /////////////////////////////////////////////////////////
@@ -43,9 +47,14 @@ if (!happyhour.count() && !etablishment.count()) {
             title: 'Attention'
         });
         dialog.show();
-    }else{
 
-        Ti.API.info("Get All data");
+        Alloy.Globals.endDownload = true;
+
+        Alloy.Collections.etablishment.fetch();
+
+        Alloy.Collections.etablishment.sort();
+
+    }else{
 
         Alloy.Globals.getAllData();
 
@@ -54,19 +63,35 @@ if (!happyhour.count() && !etablishment.count()) {
 
 }else {
     if (!Alloy.Globals.hasConnection()) {
-        var dialog = Ti.UI.createAlertDialog({
-            message: 'Afin de voir les Happy hours Toulousains, veuillez vous connecter à internet au moins une fois.',
-            ok: 'Je comprends',
-            title: 'Attention'
-        });
-        dialog.show();
+
+        Alloy.Globals.endDownload = true;
+
+        Alloy.Collections.etablishment.fetch();
+
+        Alloy.Collections.etablishment.sort();
+
     }else {
+
+        Alloy.Globals.endDownload = true;
+
+        Alloy.Collections.etablishment.fetch();
+
+        Alloy.Collections.etablishment.sort();
         // TODO : Update data after 24h
     }
 }
 
 
 function download() {
+
+    if (!Alloy.Globals.hasConnection) {
+        activityIndicator.hide();
+        chargement.close();
+
+        $.tabgroup.open();
+
+        return;
+    }
     if(Alloy.Globals.endDownload || i > 5000) {
 
         activityIndicator.hide();
